+++ tests/integration_tests/test_hasura.py
@@ -12,11 +12,9 @@ from dipdup.config import HasuraConfig
 from dipdup.config import PostgresDatabaseConfig
 from dipdup.dipdup import DipDup
 from dipdup.hasura import HasuraGateway
-from dipdup.utils import skip_ci
 from dipdup.utils.database import tortoise_wrapper
 
 
-@skip_ci
 class HasuraTest(IsolatedAsyncioTestCase):
     maxDiff = None
 
@@ -44,10 +42,10 @@ class HasuraTest(IsolatedAsyncioTestCase):
             dipdup = DipDup(config)
             await stack.enter_async_context(tortoise_wrapper(config.database.connection_string, 'demo_hic_et_nunc.models'))
             await dipdup._set_up_database(stack)
-            await dipdup._set_up_hooks()
+            await dipdup._set_up_hooks(set())
             await dipdup._initialize_schema()
 
-            hasura_container = DbContainer('hasura/graphql-engine:v2.4.0').with_env(
+            hasura_container = DbContainer('hasura/graphql-engine:v2.6.2').with_env(
                 'HASURA_GRAPHQL_DATABASE_URL',
                 f'postgres://test:test@{postgres_ip}:5432',
             )